rename Container components
