+++ src/components/SignUpContainer.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Redirect } from 'react-router-dom';
 import { Logo } from './Login/Logo';
-import classes from './SignUp.module.css';
+import classes from './SignUpContainer.module.css';
 import { EnterProfile } from './SignUp/EnterProfile';
 import { EnterUserImage } from './SignUp/EnterUserImage';
 import { EnterBio } from './SignUp/EnterBio';
@@ -18,7 +18,7 @@ type Props = {
 };
 
 // todo container と presentation にわけよう
-export const SignUp: React.FC<Props> = (props) => {
+export const SignUpContainer: React.FC<Props> = (props) => {
   const { isLogin } = props;
 
   // Common
